Add KDoc comments for all KonigDecimal functions, refactor KonigDecimal tests separate from FixedKonigDecimal tests
